refactor: Place JavaFX classes into corresponding directories.
