+++ UML editor/pom.xml
@@ -74,7 +74,7 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.8</version>
                 <configuration>
-                    <mainClass>codecain.GraphicalInterfaceJavaFX</mainClass>
+                    <mainClass>codecain.GraphicalUserInterface.GraphicalInterfaceJavaFX</mainClass>
                 </configuration>
             </plugin>
         </plugins>
