+++ UML editor/src/main/java/codecain/Main.java
@@ -1,5 +1,7 @@
 package codecain;
 
+import codecain.GraphicalUserInterface.GraphicalInterfaceJavaFX;
+
 public class Main {
     public static void main(String[] args) {
         GraphicalInterfaceJavaFX.main(args);
