--- UML editor/src/main/resources/GraphicalInterface.fxml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ToolBar?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane fx:id="classContainer" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1000.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="codecain.Controller">
-   <children>
-      <Button layoutX="814.0" layoutY="747.0" mnemonicParsing="false" onAction="#addClassBtn" text="Add Class" />
-      <Button layoutX="900.0" layoutY="747.0" mnemonicParsing="false" onAction="#deleteClassBtn" text="Delete Class" />
-      <ToolBar layoutX="130.0" layoutY="14.0" prefHeight="40.0" prefWidth="200.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-        <items>
-          <Button mnemonicParsing="false" text="Save" />
-            <Button layoutX="10.0" layoutY="12.0" mnemonicParsing="false" text="Load" />
-            <Button layoutX="68.0" layoutY="12.0" mnemonicParsing="false" text="Help" />
-            <Button layoutX="126.0" layoutY="12.0" mnemonicParsing="false" text="Exit" />
-        </items>
-      </ToolBar>
-   </children>
-</AnchorPane>
