+++ UML editor/src/main/resources/codecain/GraphicalInterface.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.control.ToolBar?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane fx:id="classContainer" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1000.0" stylesheets="@application.css" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="codecain.Controller">
+<AnchorPane fx:id="classContainer" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1000.0" stylesheets="@application.css" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="codecain.GraphicalUserInterface.Controller">
    <children>
       <AnchorPane fx:id="nodeContainer" prefHeight="800.0" prefWidth="1000.0" />
       <Button layoutX="814.0" layoutY="747.0" mnemonicParsing="false" onAction="#addClassBtn" text="Add Class" />
